Removed duplicate contract files Renamed helper files Fixed broken file paths Simplified some logic
