+++ test/block_tags.js
@@ -12,7 +12,7 @@ var to = require("../lib/utils/to.js");
 // This removes solc's overzealous uncaughtException event handler.
 process.removeAllListeners("uncaughtException");
 
-var source = fs.readFileSync("./test/Example.sol", { encoding: "utf8" });
+var source = fs.readFileSync("./test/contracts/Example.sol", { encoding: "utf8" });
 var result = solc.compile(source, 1);
 
 // Note: Certain properties of the following contract data are hardcoded to
