+++ test/forkingasprovider.js
@@ -38,7 +38,7 @@ describe("Forking using a Provider", function() {
 
   before("set up test data", function() {
     this.timeout(5000);
-    var source = fs.readFileSync("./test/Example.sol", { encoding: "utf8" });
+    var source = fs.readFileSync("./test/contracts/Example.sol", { encoding: "utf8" });
     var result = solc.compile(source, 1);
 
     // Note: Certain properties of the following contract data are hardcoded to
